Create black_deck.py
Its a two player game, one is the user and the other one is computer. both of them compete to win the game
that is to grab the money on the bet (placed by the user) if player wins the money on bet will get doubled and if looses
the money will be deducted from the account.
player has a goal to reach as close as possible to the target of 21 by hitting or standing. if the user stand than the result
will be declared, which is, which player is more closer to 21 and the awarded accordingly.